+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 
 TODO: Version 3.1.0
 --------------------------------------
-- #709: Publish Groovy 4.x-specific jar for CodeNarc. Upgrade to GMetrics 2.1.0.
-- #705: Make SourceFileTest more multi-platform friendly. ([Paul King](https://github.com/paulk-asert))
+ - #709: Publish Groovy 4.x-specific jar for CodeNarc. Upgrade to GMetrics 2.1.0.
+ - #705: Make SourceFileTest more multi-platform friendly. ([Paul King](https://github.com/paulk-asert))
+ - #688: **SpaceAfterMethodCallName**: Fix false positives. NOTE: No longer check method calls without parentheses. 
 
 
 Version 3.0.1    (Apr 2022)